re-structure the code
