+++ app/src/main/java/org/oppia/android/app/spotlight/SpotlightTarget.kt
@@ -1,0 +1,56 @@
+package org.oppia.android.app.spotlight
+
+import android.view.View
+
+data class SpotlightTarget(
+  val anchor: View,
+  val hint: String = "",
+  val shape: SpotlightShape = SpotlightShape.RoundedRectangle,
+  val feature: org.oppia.android.app.model.Spotlight.FeatureCase
+) {
+  val anchorLeft: Float = calculateAnchorLeft()
+  val anchorTop: Float = calculateAnchorTop()
+  val anchorHeight: Int = calculateAnchorHeight()
+  val anchorWidth: Int = calculateAnchorWidth()
+  val anchorCentreX = calculateAnchorCentreX()
+  val anchorCentreY = calculateAnchorCentreY()
+
+  init {
+    calculateAnchorLeft()
+    calculateAnchorTop()
+    calculateAnchorHeight()
+    calculateAnchorWidth()
+    calculateAnchorCentreX()
+    calculateAnchorCentreY()
+  }
+
+  private fun calculateAnchorLeft(): Float {
+    val location = IntArray(2)
+    anchor.getLocationOnScreen(location)
+    val x = location[0]
+    return x.toFloat()
+  }
+
+  private fun calculateAnchorTop(): Float {
+    val location = IntArray(2)
+    anchor.getLocationOnScreen(location)
+    val y = location[1]
+    return y.toFloat()
+  }
+
+  private fun calculateAnchorHeight(): Int {
+    return anchor.height
+  }
+
+  private fun calculateAnchorWidth(): Int {
+    return anchor.width
+  }
+
+  private fun calculateAnchorCentreX(): Float {
+    return anchorLeft + anchorWidth / 2
+  }
+
+  private fun calculateAnchorCentreY(): Float {
+    return anchorTop + anchorHeight / 2
+  }
+}
